import fixed, package built, readme updated
