+++ Battleships_api/Ai.py
@@ -1,4 +1,4 @@
-import References
+from . import References
 import random
 
 class Ai():
