+++ issues/current_issues.md
@@ -1,0 +1,1 @@
+Open issues will be listed here, along with the person that has taken ownership of the issue. Either email me on 55702741+simalpas@users.noreply.github.com or fork and add to this directly with a pull request.
