+++ setup.cfg
@@ -1,8 +1,8 @@
 [metadata]
 name = Battleships_api
-version = 0.0.1
+version = 0.1.0
 author = Simon Malpas
-author_email = 
+author_email = 55702741+simalpas@users.noreply.github.com
 description = Play Battleships, incudes a basic cli front end
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -16,9 +16,9 @@ classifiers =
 
 [options]
 package_dir =
-    = src
+    = Battleships_api
 packages = find:
 python_requires = >=3.6
 
 [options.packages.find]
-where = src
+where = Battli Battleships_api
