--- src/Battleships_api/__init__.py
@@ -1,5 +1,0 @@
-from Battleships import Battleships
-from GameBoard import GameBoard
-from Player import Player
-from Ai import Ai
-import References
